backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
chromadb==0.6.3
openai==1.59.5
python-dotenv==1.0.1
python-multipart==0.0.20
pypdf==5.1.0
requests==2.32.3
pydantic==2.10.4
httpx==0.28.1
